+++ src/main.ts
@@ -1,10 +1,12 @@
 import { Plugin } from "obsidian";
+import { inlineAnnotationLivePreviewExtension } from "./live-preview";
 import { renderInlineAnnotationsInElement } from "./postprocessor";
 
 export default class InlineAnnotationPlugin extends Plugin {
-  async onload() {
+  onload() {
     this.registerMarkdownPostProcessor((element) => {
       renderInlineAnnotationsInElement(element);
     });
+    this.registerEditorExtension(inlineAnnotationLivePreviewExtension());
   }
 }