+++ test/run-tests.ts
@@ -1,0 +1,2 @@
+import "./postprocessor.test";
+import "./live-preview-ranges.test";